+++ CHANGELOG.md
@@ -41,6 +41,7 @@
 * Get rid of outdated componentWillReceiveProps method in SearchAndSort component (UIDATIMP-1001)
 * Cover `<EditKeyShortcutsWrapper` component with tests (UIDATIMP-957)
 * refactor psets away from backend ".all" permissions (UIDATIMP-1017)
+* Cover `<MatchingFieldsManager>` component with tests (UIDATIMP-712)
 
 ### Bugs fixed:
 * Auxiliary "repeatableFieldAction" property disappear while removing "vendor reference number" field mapping (UIDATIMP-987)