+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## **5.1.0** (in progress)
 
+### Features added:
+* Cover `<MatchingFieldsManager>` component with tests (UIDATIMP-712)
+* Cover `<EditKeyShortcutsWrapper` component with tests (UIDATIMP-957)
+
 ## [5.0.0](https://github.com/folio-org/ui-data-import/tree/v5.0.0) (2021-10-08)
 
 ### Features added: