+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features added:
 * Cover `<MatchingFieldsManager>` component with tests (UIDATIMP-712)
 * Cover `<EditKeyShortcutsWrapper` component with tests (UIDATIMP-957)
+* Cover `<ViewAllLogs>` component with tests (UIDATIMP-969)
 * Cover `<ReturnToAssignJobs>` component with tests (UIDATIMP-960)
 
 ## [5.0.0](https://github.com/folio-org/ui-data-import/tree/v5.0.0) (2021-10-08)
@@ -45,6 +46,8 @@
 * Cover `<DetailsKeyShortcutsWrapper>` component with tests (UIDATIMP-956)
 * Get rid of outdated componentWillReceiveProps method in SearchAndSort component (UIDATIMP-1001)
 * refactor psets away from backend ".all" permissions (UIDATIMP-1017)
+* Cover `<MatchingFieldsManager>` component with tests (UIDATIMP-712)
+* Cover `<FieldOrganization>` component with tests (UIDATIMP-958)
 
 ### Bugs fixed:
 * Auxiliary "repeatableFieldAction" property disappear while removing "vendor reference number" field mapping (UIDATIMP-987)