+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 ## **5.1.0** (in progress)
 
+### Features added:
+* Cover `<MatchingFieldsManager>` component with tests (UIDATIMP-712)
+* Cover `<EditKeyShortcutsWrapper` component with tests (UIDATIMP-957)
+* Cover `<ViewAllLogs>` component with tests (UIDATIMP-969)
+* Cover `<ReturnToAssignJobs>` component with tests (UIDATIMP-960)
+* Cover `<MatchingFieldsManager>` component with tests (UIDATIMP-712)
+* Cover `<FieldOrganization>` component with tests (UIDATIMP-958)
+* Add AUTHORITY type into folioRecordTypes (UIDATIMP-1021)
+
+### Bugs fixed:
+* Illegal character in query for mod-source-record-manager (UIDATIMP-1004)
+
 ## [5.0.0](https://github.com/folio-org/ui-data-import/tree/v5.0.0) (2021-10-08)
 
 ### Features added: